api/src/requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
django-registration-redux==2.13
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-allauth==0.58.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary
requests
django-filter
django-pluggableappsettings
django-rest-swagger
drf-yasg
redis
google-cloud-bigquery
google-cloud-secret-manager
randomname
pandas